pyproject.toml
[project]
name = "nt-data-pipelines"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alpaca-py>=0.43.2",
    "bear-lake>=0.1.5",
    "beautifulsoup4>=4.14.3",
    "clickhouse-connect>=0.10.0",
    "cvxpy>=1.7.5",
    "lxml>=6.0.2",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pandas-market-calendars>=5.2.3",
    "polars>=1.36.1",
    "prefect>=3.6.8",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "ray[default]>=2.53.0",
    "requests>=2.32.5",
    "statsmodels>=0.14.6",
    "tqdm>=4.67.1",
    "yfinance>=0.2.50",
]

[dependency-groups]
dev = [
    "matplotlib>=3.10.8",
    "rich>=14.2.0",
    "seaborn>=0.13.2",
]
